Add Matrix Comminuty
I have made two rooms on Matrix one in English and another in Persian so that those who care to be on Ethical and FLOSS communication platform have some place to share thoughts and experience. I have some degree of knowledge on Alpine and has implemented few projects. I would welcome more experienced people to join the rooms. I would welcome connecting the room to other communities through Matrix Bridge. That is up to the admins of other communities.