remove: Telegram community
We've made the decision to remove the Telegram community link from the repository. There has been quite a bit of toxic behaviour and it's not very welcoming to people who are seeking help.

If you would like to join one of the other communities, please feel free. We're all very active on Discord, as well as Matrix.